problemas com webpack

+++ src/Pages/Admin.js
@@ -5,9 +5,10 @@ import { faEdit,faTrashAlt } from '@fortawesome/free-solid-svg-icons'
 import NavBar from '../Components/NavBar'
 import axios from 'axios'
 import EditUser from '../Components/EditUser'
-import { Modal, ModalHeader, ModalBody, ModalFooter } from 'reactstrap';
+import { Modal, ModalHeader, ModalBody, ModalFooter,NavItem,NavLink } from 'reactstrap';
 import SignUp from './SignUp'
-
+import swal from 'sweetalert';
+import { Link } from 'react-router-dom'
 
 const Admin=(props)=>{
 
@@ -19,7 +20,8 @@ const Admin=(props)=>{
     const [modal, setModal] = useState(false);
     
     const toggle = () => setModal(!modal);
-
+    
+    const[userChanges,setUserChanges]=useState(false);
     const[users,setUsers]=useState([]);
     const[user,setUser]=useState();
 
@@ -28,27 +30,51 @@ const Admin=(props)=>{
     }).then(res=>{
         setUsers(res.data)
     })
-    },[])
+    },[userChanges])
 
     const handleDelete=(id)=>{
-        console.log(id);
-        axios.delete(`http://localhost:8080/api/v1/users/${id}`,{headers:{"authorization":localStorage.getItem("@TOKEN")}}).then(res=>{
-            console.log(res);
-        })
+
+        swal({
+            title: "Tem certeza?",
+            text:"Deseja mesmo apagar esse usuário?",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+          })
+          .then((willDelete) => {
+            if (willDelete) {
+                axios.delete(`http://localhost:8080/api/v1/users/${id}`,{headers:{"authorization":localStorage.getItem("@TOKEN")}}).then(res=>{
+                   if(res.status===200){
+                    swal("usuario deletado", {
+                        icon: "success",
+                      });
+                      setUserChanges(!userChanges);
+                   }
+                })
+           
+            } else {
+              swal("Operação cancelada");
+            }
+          });
     }
 
     const handleSubmit=(e)=>{
-        console.log(user);
         e.preventDefault();
         axios.post(`http://localhost:8080/api/v1/users`,{username:user.username,password:user.password},{headers:{"authorization":localStorage.getItem("@TOKEN")}})
         .then(res=>{
-            console.log(res);
+            if(res.status===201){
+                swal("Usuário Criado com Sucesso!", {
+                    icon: "success",
+                  }).then(res=>{
+                    toggle();
+                    setUserChanges(!userChanges);
+                })
+            }
         })
     }
 
     return(
         <>
-        <NavBar/>
         <Container>
             <Table hover>
                 <thead>
@@ -72,7 +98,7 @@ const Admin=(props)=>{
                             <td>{user.role.name}</td>) :
                             (<td>Null</td>)}
                             
-                            <td ><EditUser data={user}/></td>
+                            <td ><EditUser data={user} userChanges={userChanges} setUserChanges={setUserChanges}/></td>
                             <td ><FontAwesomeIcon icon={faTrashAlt} className="delete-icon" style={{cursor:"pointer"}} onClick={()=>handleDelete(user.id)}/></td>
                         </tr>
                         )
